chess_minmax/requirements.txt
CairoSVG == 2.7.1
gym == 0.17.3
gym_chess == 0.1.1
imageio == 2.37.0
python_chess == 0.31.4
imageio[ffmpeg]
setuptools